WIP working on unfolding images : 	- separate utils files to make it clearer 	- images unfolded, but not plotted yet
